pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "musicclassifier"
version = "0.1.0"
description = "QQ音乐歌单自动处理器 - 获取、分类、管理和分析你的QQ音乐歌单"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
authors = [
    { name = "Your Name", email = "you@example.com" },
]

dependencies = [
    "httpx>=0.27",
    "pandas>=2.0",
    "typer>=0.12",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "pyyaml>=6.0",
    "loguru>=0.7",
    "openpyxl>=3.1",
    "rich>=13.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.4",
]

[project.scripts]
musicclassifier = "musicclassifier.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/musicclassifier"]

[tool.ruff]
target-version = "py310"
line-length = 100

[tool.pytest.ini_options]
testpaths = ["tests"]
